organized file structure, added view class, working on help, AB#198
