--- src/main/java/com/destination/adventure/Destination.java
@@ -1,5 +1,0 @@
-package com.destination.adventure;
-
-public class Destination {
-
-}
