+++ src/main/java/com/destination/adventure/model/Destination.java
@@ -1,0 +1,5 @@
+package com.destination.adventure.model;
+
+public class Destination {
+
+}
